Move initializer into separate module
