+++ lib/main.js
@@ -52,52 +52,15 @@ import './animations/poplist-open-default';
 import './animations/poplist-close-flyin';
 import './animations/poplist-open-flyin';
 
+import EuiInitializer from './initializers/eui-initializer';
 
-Ember.Application.initializer({
-  name: 'emberui',
 
-  initialize: function(container) {
-    container.register('template:components/eui-button', EuiButtonTemplate);
-    container.register('component:eui-button', EuiButtonComponent);
-
-    container.register('template:components/eui-checkbox', EuiCheckboxTemplate);
-    container.register('component:eui-checkbox', EuiCheckboxComponent);
-
-    container.register('template:components/eui-dropbutton', EuiDropbuttonTemplate);
-    container.register('component:eui-dropbutton', EuiDropbuttonComponent);
-
-    container.register('template:components/eui-input', EuiInputTemplate);
-    container.register('component:eui-input', EuiInputComponent);
-
-    container.register('template:components/eui-modal', EuiModalTemplate);
-    container.register('component:eui-modal', EuiModalComponent);
-
-    container.register('template:components/eui-poplist', EuiPoplistTemplate);
-    container.register('template:components/eui-poplist-opion', EuiPoplistOptionTemplate);
-    container.register('component:eui-poplist', EuiPoplistComponent);
-
-    container.register('template:components/eui-select', EuiSelectTemplate);
-    container.register('component:eui-select', EuiSelectComponent);
-
-    container.register('template:components/eui-selectdate', EuiSelectDateTemplate);
-    container.register('component:eui-selectdate', EuiSelectDateComponent);
-
-    container.register('template:components/eui-popcal', EuiPopcalTemplate);
-    container.register('component:eui-popcal', EuiPopcalComponent);
-
-    container.register('template:components/eui-textarea', EuiTextareaTemplate);
-    container.register('component:eui-textarea', EuiTextareaComponent);
-
-    container.register('component:eui-month', EuiMonthComponent);
-
-    container.register('template:components/eui-calendar', EuiCalendarTemplate);
-    container.register('component:eui-calendar', EuiCalendarComponent);
-  }
-});
+Ember.Application.initializer(EuiInitializer);
 
 Ember.libraries.register('EmberUI', '0.1.2');
 
 export {
+  EuiInitializer,
   EuiButtonComponent,
   EuiCheckboxComponent,
   EuiDropbuttonComponent,
